requirements.txt
# Python Requirements
# Copyright (c) 2022 Cannabis Data
# Copyright (c) 2021-2022 Cannlytics
# Created 1/5/2021
# Updated: 11/22/2022
cannlytics==0.0.12
datetime==4.7
django==4.1.3
django-environ==0.9.0
django-feather==0.3.0
django-livereload-server==0.4
django-permissions-policy==4.13.0
django-robohash-svg==0.9.5
djangorestframework==3.14.0
firebase_admin==6.0.1
fredapi==0.5.0
google-auth==2.14.1
google-cloud-secret-manager==2.12.6
google-cloud-firestore==2.7.2
googlemaps==4.7.3
gunicorn==20.1.0
Markdown==3.4.1
openpyxl==3.0.10
opencv-python==4.6.0.66
pandas==1.5.1
pdfplumber==0.7.6
python-dateutil==2.8.2
python-dotenv==0.21.0
pymdown-extensions==9.8
python-markdown-math==0.8
pyzbar==0.1.9
scikit-learn==1.1.3
selenium==4.6.0
spacy==3.4.3
statsmodels==0.13.5
ulid-py==1.1.0
waitress==2.1.2
whitenoise==6.2.0
zipcodes==1.2.0
